requirements.txt
Flask==1.1.4
pydantic==1.9.0
google-api-python-client
requests==2.26.0
google-generativeai
Werkzeug==1.0.0
MarkupSafe==2.0.1
sec-edgar-downloader==5.0.2